source code organization; things should be a lot cleaner now
